shower sfx hotfix

+++ Core/Map/Tiles.cs
@@ -246,7 +246,7 @@ public class TileShower : Tile, IInteractable
     {
         if (e is Player p)
         {
-            showerSoundInstance.Volume = (float)Settings.SfxVolume.Value;
+            showerSoundInstance.Volume = (float)(double)Settings.SfxVolume.Value;
             showerSoundInstance.Play();
 
             var existingEffect = p.Inventory.Effects.FirstOrDefault(effect => effect is EffectLol);
